tests: ensure expected error messages for negative tests
Prior to this patch, the test suite simply checked if negative tests
failed.  However, such failures are ambiguous since we don't check the
precise reason for the failure.  This patch adds the expected error
messages.  As a result of this patch, all new negative tests will need
to be explicitly listed in the tests.ml file, along with the expected
error message.