shdr: add code for parsing section headers
This patch also adds the code to handle the case when the section header
count is stored in the size field of the first section header, although
I don't know of a good way to test this.  Do I need to create an ELF
file with more than 2^16 section headers?